Fixed Issue #68 (Padding preserves transparency)
Previously, a new buffered image was created (with the larger
dimensions). Next, the entire image was filled with the border color.
Finally, the original image was drawn on top of the new image. This
created issue #68 where transparency was lost because the background was
being filled with a color.

Now, a new buffered image is created (with the larger dimensions). Next,
only the padded parts of the image are drawn in. Finally, the old image
is drawn on top of the new image. By not messing with the background
behind the old image, transparency is preserved.