Fixed automatic resizing for non-square targets.
Previously, automatic mode only respected the longer dimension of the source image. That would lead to an improper scaled image, if the aspect ratio (w/h) for landscape image of the source is smaller than the aspect ratio of the target size. 

For example, a nearly square but still landscape source image resized to a targe size of 200px x 150px resulted in an image height larger than 150px.